refactor(ffmpeg): split LocalAdapter god-file by concern (1459→287)
adapter.go bundled four unrelated concerns on LocalAdapter. Split into
cohesive sibling files within package ffmpeg — pure relocation, no
call-site or behavior changes:

- adapter.go (287): struct, ctor, shared types/consts/vars
- adapter_benchmark.go: preflight + profile/encoder benchmarking,
  benchmark filters, hardware-codec queries
- adapter_logparse.go: ffmpeg stderr/log scanning, value parsing,
  failure-line classification, runtime diagnostics
- adapter_process.go: process lifecycle (Start/Stop/monitor/Health),
  group termination, env bounds, process-detail bookkeeping

53 funcs in, 53 funcs out (lossless). Verified by go build, go vet,
the ffmpeg argv/golden + package tests, and scoped golangci-lint;
config surfaces unchanged.
