+++ backend/internal/infra/media/ffmpeg/adapter.go
@@ -1,39 +1,19 @@
 package ffmpeg
 
 import (
-	"bufio"
-	"bytes"
 	"context"
-	"errors"
-	"fmt"
-	"io"
-	"net"
-	"net/http"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strconv"
-	"strings"
-	"sync"
-	"sync/atomic"
-	"time"
-
 	"github.com/ManuGH/xg2g/internal/config"
 	playbackports "github.com/ManuGH/xg2g/internal/domain/playbackprofile/ports"
 	"github.com/ManuGH/xg2g/internal/domain/session/ports"
 	"github.com/ManuGH/xg2g/internal/domain/vod"
-	"github.com/ManuGH/xg2g/internal/infra/ffmpeg/watchdog"
-	"github.com/ManuGH/xg2g/internal/infra/media/ffmpeg/capability"
-	"github.com/ManuGH/xg2g/internal/metrics"
 	"github.com/ManuGH/xg2g/internal/pipeline/exec/enigma2"
-	"github.com/ManuGH/xg2g/internal/pipeline/hardware"
-	"github.com/ManuGH/xg2g/internal/pipeline/profiles"
-	"github.com/ManuGH/xg2g/internal/procgroup"
-	"github.com/ManuGH/xg2g/internal/telemetry"
 	"github.com/rs/zerolog"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/trace"
+	"net"
+	"net/http"
+	"os/exec"
+	"strings"
+	"sync"
+	"time"
 )
 
 const (
@@ -305,1155 +285,3 @@ func NewLocalAdapter(binPath string, ffprobeBin string, hlsRoot string, e2 *enig
 	adapter.detector.terminateProcessGroup = adapter.terminateProcessGroup
 	return adapter
 }
-
-func profileBenchmarksForBackend(backend string) []string {
-	switch backend {
-	case "cpu":
-		return startupProfilesToBenchmark
-	case "vaapi", "nvenc":
-		return []string{
-			playbackports.BenchmarkProfileVideoH2641080P,
-			playbackports.BenchmarkProfileVideoH2641080I,
-			playbackports.BenchmarkProfileVideoH2641080I50,
-			playbackports.BenchmarkProfileVideoH2642160P,
-			playbackports.BenchmarkProfileVideoH2642160P50,
-		}
-	default:
-		return nil
-	}
-}
-
-func vaapiEncodeOnlyInterlacedCorrectnessFilter(encoder string) string {
-	parts := []string{
-		"setfield=tff",
-		"bwdif=mode=send_field:parity=auto:deint=all",
-	}
-	if normalizeRequestedCodec(encoder) == "av1" {
-		parts = append(parts, av1VAAPIGeometryPadFilter())
-	}
-	parts = append(parts, "format=nv12", "hwupload")
-	return strings.Join(parts, ",")
-}
-
-func isAV1SignalStatsDecodeUnavailable(err error) bool {
-	if err == nil {
-		return false
-	}
-	msg := strings.ToLower(err.Error())
-	return strings.Contains(msg, "doesn't support hardware accelerated av1 decoding") ||
-		strings.Contains(msg, "error submitting packet to decoder")
-}
-
-func outputFileHasBytes(path string) bool {
-	info, err := os.Stat(path)
-	return err == nil && info.Size() > 0
-}
-
-func cpuProfileBenchmarkFilter(profileID string) string {
-	switch strings.ToLower(strings.TrimSpace(profileID)) {
-	case playbackports.BenchmarkProfileVideoH2641080I, playbackports.BenchmarkProfileVideoH2641080I50:
-		return "setfield=tff,bwdif=mode=send_field:parity=auto:deint=all"
-	default:
-		return ""
-	}
-}
-
-func vaapiProfileBenchmarkFilter(profileID string) (string, error) {
-	switch strings.ToLower(strings.TrimSpace(profileID)) {
-	case playbackports.BenchmarkProfileVideoH2641080P:
-		return "format=nv12,hwupload", nil
-	case playbackports.BenchmarkProfileVideoH2641080I50:
-		return "format=nv12,setfield=tff,hwupload,deinterlace_vaapi", nil
-	case playbackports.BenchmarkProfileVideoH2642160P:
-		return "format=nv12,hwupload", nil
-	case playbackports.BenchmarkProfileVideoH2642160P50:
-		return "format=nv12,hwupload", nil
-	case playbackports.BenchmarkProfileVideoH2641080I:
-		return "format=nv12,setfield=tff,hwupload,deinterlace_vaapi", nil
-	default:
-		return "", fmt.Errorf("unsupported vaapi benchmark profile %q", profileID)
-	}
-}
-
-func nvencProfileBenchmarkFilter(profileID string) string {
-	switch strings.ToLower(strings.TrimSpace(profileID)) {
-	case playbackports.BenchmarkProfileVideoH2641080I, playbackports.BenchmarkProfileVideoH2641080I50:
-		return "setfield=tff,bwdif=mode=send_field:parity=auto:deint=all"
-	default:
-		return ""
-	}
-}
-
-func profileBenchmarkInput(profileID string) string {
-	switch strings.ToLower(strings.TrimSpace(profileID)) {
-	case playbackports.BenchmarkProfileVideoH2642160P50:
-		return "testsrc=duration=0.2:size=3840x2160:rate=50"
-	case playbackports.BenchmarkProfileVideoH2642160P:
-		return "testsrc=duration=0.2:size=3840x2160:rate=25"
-	case playbackports.BenchmarkProfileVideoH2641080I50:
-		return "testsrc=duration=0.2:size=1920x1080:rate=50"
-	default:
-		return "testsrc=duration=0.2:size=1920x1080:rate=25"
-	}
-}
-
-func profileBenchmarkTimeout(profileID string) time.Duration {
-	switch strings.ToLower(strings.TrimSpace(profileID)) {
-	case playbackports.BenchmarkProfileVideoH2642160P50:
-		return 22 * time.Second
-	case playbackports.BenchmarkProfileVideoH2642160P:
-		return 18 * time.Second
-	case playbackports.BenchmarkProfileVideoH2641080I50:
-		return 15 * time.Second
-	default:
-		return 12 * time.Second
-	}
-}
-
-func runProfileBenchmarkCommand(ctx context.Context, binPath string, args []string) (time.Duration, error) {
-	start := time.Now()
-	// #nosec G204 -- BinPath is trusted from config and args are fixed synthetic probes.
-	cmd := exec.CommandContext(ctx, binPath, args...)
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return 0, fmt.Errorf("profile benchmark failed: %w (output: %s)", err, string(out))
-	}
-	return time.Since(start), nil
-}
-
-func (a *LocalAdapter) supportedHWCodecsLocal() []string {
-	codecs := make([]string, 0, 3)
-	for _, codec := range []string{"h264", "hevc", "av1"} {
-		if a.detector.preferredHardwareBackendForCodec(codec) != profiles.GPUBackendNone {
-			codecs = append(codecs, codec)
-		}
-	}
-	return codecs
-}
-
-func (a *LocalAdapter) autoHWCodecsLocal() []string {
-	codecs := make([]string, 0, 3)
-	for _, codec := range []string{"h264", "hevc", "av1"} {
-		for _, backend := range []profiles.GPUBackend{profiles.GPUBackendVAAPI, profiles.GPUBackendNVENC} {
-			encoder, ok := capability.EncoderNameForBackend(codec, backend)
-			if !ok {
-				continue
-			}
-			if a.detector.hardwareEncoderAutoEligible(backend, encoder) {
-				codecs = append(codecs, codec)
-				break
-			}
-		}
-	}
-	return codecs
-}
-
-// Start initiates the media process.
-func (a *LocalAdapter) Start(ctx context.Context, spec ports.StreamSpec) (ports.RunHandle, error) {
-	if spec.Source.Type == ports.SourceTuner && a.E2 != nil {
-		if spec.Source.TunerSlot < 0 {
-			return "", fmt.Errorf("invalid tuner slot: %d", spec.Source.TunerSlot)
-		}
-		tuner := enigma2.NewTuner(a.E2, spec.Source.TunerSlot, 10*time.Second)
-		if err := tuner.Tune(ctx, spec.Source.ID); err != nil {
-			return "", fmt.Errorf("tuning failed: %w", err)
-		}
-		a.Logger.Info().
-			Str("session_id", spec.SessionID).
-			Str("startup_phase", "tuner_tuned").
-			Int("tuner_slot", spec.Source.TunerSlot).
-			Str("service_ref", spec.Source.ID).
-			Msg("tuner tune completed")
-	}
-
-	inputURL := ""
-	switch spec.Source.Type {
-	case ports.SourceTuner:
-		if a.E2 == nil {
-			return "", fmt.Errorf("tuner source requires enigma2 client")
-		}
-		streamURL, err := a.E2.ResolveStreamURL(ctx, spec.Source.ID)
-		if err != nil {
-			return "", fmt.Errorf("resolve stream url: %w", err)
-		}
-		streamURL = a.injectCredentialsIfAllowed(streamURL)
-		a.Logger.Info().
-			Str("session_id", spec.SessionID).
-			Str("startup_phase", "stream_url_resolved").
-			Str("resolved_url", sanitizeURLForLog(streamURL)).
-			Msg("stream url resolved")
-
-		chosenURL, err := a.selectStreamURL(ctx, spec.SessionID, spec.Source.ID, streamURL)
-		if err != nil {
-			return "", err
-		}
-		inputURL = chosenURL
-		a.Logger.Info().
-			Str("session_id", spec.SessionID).
-			Str("startup_phase", "input_url_selected").
-			Str("input_url", sanitizeURLForLog(inputURL)).
-			Msg("stream input url selected")
-	case ports.SourceURL:
-		inputURL = spec.Source.ID
-		a.Logger.Info().
-			Str("session_id", spec.SessionID).
-			Str("startup_phase", "input_url_selected").
-			Str("input_url", sanitizeURLForLog(inputURL)).
-			Msg("direct input url selected")
-	}
-	sourceKey := fpsCacheKey(spec.Source, inputURL)
-
-	a.Logger.Info().
-		Str("session_id", spec.SessionID).
-		Str("startup_phase", "ffmpeg_args_build_started").
-		Str("input_url", sanitizeURLForLog(inputURL)).
-		Msg("ffmpeg args build started")
-	// Span covering the playback decision + ffprobe probes (the untraced gap
-	// between request and spawn). No-op when tracing is off. planCtx lets any
-	// probe spans added later nest under this one.
-	planCtx, planSpan := telemetry.Tracer("xg2g.ffmpeg").Start(ctx, "playback.plan",
-		trace.WithAttributes(
-			attribute.String("xg2g.session_id", spec.SessionID),
-			attribute.String("xg2g.source_type", fmt.Sprintf("%v", spec.Source.Type)),
-		),
-	)
-	plan, err := a.buildArgsWithPlan(planCtx, spec, inputURL)
-	if err != nil {
-		planSpan.RecordError(err)
-		planSpan.SetStatus(codes.Error, "plan build failed")
-		planSpan.End()
-		return "", fmt.Errorf("failed to build args: %w", err)
-	}
-	planSpan.SetAttributes(attribute.String("xg2g.path_id", plan.pathID))
-	planSpan.End()
-	args := plan.args
-	a.Logger.Info().
-		Str("session_id", spec.SessionID).
-		Str("startup_phase", "ffmpeg_args_built").
-		Int("arg_count", len(args)).
-		Msg("ffmpeg args build finished")
-
-	// #nosec G204 - BinPath is trusted from config; args are generated by strict internal logic (buildArgs)
-	cmd := exec.CommandContext(ctx, a.BinPath, args...)
-	procgroup.Set(cmd) // Mandatory for tree reaping
-	// On ctx cancellation, gracefully SIGTERM the whole process group so ffmpeg
-	// can flush its final segment + ENDLIST, instead of the stdlib default which
-	// SIGKILLs the leader PID only. WaitDelay then bounds how long Wait blocks
-	// after cancellation before the runtime force-kills the leader and closes I/O.
-	cmd.Cancel = func() error {
-		if cmd.Process == nil {
-			return nil
-		}
-		return procgroup.TerminateGroup(cmd.Process.Pid)
-	}
-	if a.KillTimeout > 0 {
-		cmd.WaitDelay = a.KillTimeout
-	} else {
-		cmd.WaitDelay = 5 * time.Second
-	}
-
-	stderr, err := cmd.StderrPipe()
-	if err != nil {
-		return "", fmt.Errorf("failed to pipe stderr: %w", err)
-	}
-	cmd.Stdout = nil
-
-	// Span covering ffmpeg spawn -> first HLS segment (the user-perceived startup
-	// latency). No-op when tracing is disabled; ended exactly once in the monitor.
-	spawnedAt := time.Now()
-	_, startupSpan := telemetry.Tracer("xg2g.ffmpeg").Start(ctx, "ffmpeg.startup",
-		trace.WithAttributes(
-			attribute.String("xg2g.session_id", spec.SessionID),
-			attribute.String("xg2g.source_type", fmt.Sprintf("%v", spec.Source.Type)),
-			attribute.String("xg2g.hw_backend", fmt.Sprintf("%v", argsHardwareBackend(args))),
-			attribute.String("xg2g.path_id", plan.pathID),
-		),
-	)
-
-	if err := cmd.Start(); err != nil {
-		startupSpan.RecordError(err)
-		startupSpan.SetStatus(codes.Error, "ffmpeg start failed")
-		startupSpan.End()
-		return "", fmt.Errorf("ffmpeg start failed: %w", err)
-	}
-
-	pid := cmd.Process.Pid
-	a.Logger.Info().
-		Str("session_id", spec.SessionID).
-		Str("startup_phase", "ffmpeg_started").
-		Int("pid", pid).
-		Msg("ffmpeg process started")
-	handle := ports.RunHandle(fmt.Sprintf("%s-%d", spec.SessionID, pid))
-	a.mu.Lock()
-	a.activeProcs[handle] = cmd
-	a.finalizedProfiles[handle] = plan.effectiveProfile
-	// Execution truth: capture the plan parsed from the REAL argv we just handed
-	// to the process, so observers report what ffmpeg runs, not a prediction.
-	a.executedPlans[handle] = executedFFmpegPlanFromArgs(args)
-	delete(a.processDetails, handle)
-	a.mu.Unlock()
-
-	go a.monitorProcessWithStartTimeout(ctx, handle, cmd, stderr, spec.SessionID, argsHardwareBackend(args), plan.pathID, a.startTimeoutForProfile(spec.Source.Type, plan.effectiveProfile), startupSpan, spawnedAt) // #nosec G118 -- goroutine receives the request-scoped ctx (first arg), not context.Background/TODO
-	if sourceKey != "" {
-		go a.learnFPSFromOutput(sourceKey, spec.SessionID)
-	}
-
-	metrics.RecordPipelineSpawn("ffmpeg", "admitted")
-	return handle, nil
-}
-
-func (a *LocalAdapter) FinalizedProfile(handle ports.RunHandle) (ports.ProfileSpec, bool) {
-	a.mu.Lock()
-	defer a.mu.Unlock()
-
-	profile, ok := a.finalizedProfiles[handle]
-	if !ok {
-		return ports.ProfileSpec{}, false
-	}
-	return profile, true
-}
-
-// ExecutedFFmpegPlan returns the execution-truth plan parsed from the real argv
-// of the process launched for the handle. It is the un-lie-able source for
-// "what ffmpeg runs", as opposed to any profile-derived prediction.
-func (a *LocalAdapter) ExecutedFFmpegPlan(handle ports.RunHandle) (ports.ExecutedFFmpegPlan, bool) {
-	a.mu.Lock()
-	defer a.mu.Unlock()
-
-	plan, ok := a.executedPlans[handle]
-	if !ok {
-		return ports.ExecutedFFmpegPlan{}, false
-	}
-	return plan, true
-}
-
-func (a *LocalAdapter) monitorProcessWithStartTimeout(parentCtx context.Context, handle ports.RunHandle, cmd *exec.Cmd, stderr io.ReadCloser, sessionID string, hwBackend profiles.GPUBackend, pathID string, startTimeout time.Duration, startupSpan trace.Span, spawnedAt time.Time) {
-	defer func() {
-		a.mu.Lock()
-		a.removeActiveProcessLocked(handle, true)
-		a.mu.Unlock()
-	}()
-
-	// End the startup span exactly once: success is recorded on first segment;
-	// any other exit path (start timeout, early ffmpeg exit, ctx cancel) is an error.
-	var sawFirstSegment atomic.Bool
-	endStartupSpan := sync.OnceFunc(func() { startupSpan.End() })
-	defer func() {
-		if !sawFirstSegment.Load() {
-			startupSpan.SetStatus(codes.Error, "ffmpeg exited before first HLS segment")
-		}
-		endStartupSpan()
-	}()
-
-	wd := watchdog.New(startTimeout, a.StallTimeout)
-
-	if parentCtx == nil {
-		parentCtx = context.Background()
-	}
-	wdCtx, wdCancel := context.WithCancel(parentCtx)
-	defer wdCancel()
-	observerCtx, observerCancel := context.WithCancel(parentCtx)
-	defer observerCancel()
-
-	wdErrCh := make(chan error, 1)
-	go func() {
-		wdErrCh <- wd.Run(wdCtx)
-	}()
-
-	runtimeFailureLine := ""
-	scanDone := make(chan struct{})
-	go func() {
-		defer close(scanDone)
-
-		// Forward lines to RingBuffer/Log and Watchdog.
-		scanner := bufio.NewScanner(stderr)
-		scanner.Split(scanFFmpegLogTokens)
-		firstFrameLogged := false
-		firstSegmentLogged := false
-		outputObserverStarted := false
-
-		for scanner.Scan() {
-			line := scanner.Text()
-			if !firstFrameLogged {
-				if frame, ok := parseFFmpegFrameCount(line); ok && frame > 0 {
-					firstFrameLogged = true
-					wd.ObserveProgress()
-					a.writeFirstFrameMarker(sessionID)
-					a.Logger.Info().
-						Str("session_id", sessionID).
-						Str("startup_phase", "first_frame").
-						Int("frame", frame).
-						Msg("ffmpeg first frame observed")
-				}
-			}
-			if !firstSegmentLogged {
-				if segmentPath, ok := extractStartupSegmentPath(line); ok {
-					firstSegmentLogged = true
-					sawFirstSegment.Store(true)
-					startupSpan.SetAttributes(attribute.Int64("xg2g.time_to_first_segment_ms", time.Since(spawnedAt).Milliseconds()))
-					startupSpan.SetStatus(codes.Ok, "")
-					endStartupSpan()
-					wd.ObserveProgress()
-					a.Logger.Info().
-						Str("session_id", sessionID).
-						Str("startup_phase", "first_segment_write").
-						Str("segment_path", segmentPath).
-						Msg("ffmpeg first segment write observed")
-					if pathID != "" && !outputObserverStarted {
-						outputObserverStarted = true
-						go a.detector.observeRuntimePathCorrectness(observerCtx, handle, cmd, sessionID, pathID)
-					}
-				}
-			}
-			sanitizedLine := sanitizeFFmpegLogLine(line)
-			a.recordRuntimeDiagnostics(handle, line, sanitizedLine)
-			if detail := summarizeFFmpegFailureLine(sanitizedLine); detail != "" {
-				a.recordProcessDetail(handle, detail)
-			}
-			if runtimeFailureLine == "" {
-				switch hwBackend {
-				case profiles.GPUBackendVAAPI:
-					if isVAAPIRuntimeFailureLine(sanitizedLine) {
-						runtimeFailureLine = sanitizedLine
-					}
-				case profiles.GPUBackendNVENC:
-					if isNVENCRuntimeFailureLine(sanitizedLine) {
-						runtimeFailureLine = sanitizedLine
-					}
-				}
-			}
-			switch ffmpegLogLevel(sanitizedLine) {
-			case zerolog.WarnLevel:
-				a.Logger.Warn().Str("sessionId", sessionID).Str("ffmpeg_log", sanitizedLine).Msg("ffmpeg output")
-			case zerolog.InfoLevel:
-				a.Logger.Info().Str("sessionId", sessionID).Str("ffmpeg_log", sanitizedLine).Msg("ffmpeg output")
-			default:
-				a.Logger.Debug().Str("sessionId", sessionID).Str("ffmpeg_log", sanitizedLine).Msg("ffmpeg output")
-			}
-			wd.ParseLine(line)
-		}
-		if scanErr := scanner.Err(); scanErr != nil {
-			a.Logger.Warn().Err(scanErr).Str("sessionId", sessionID).Msg("ffmpeg stderr scan failed")
-		}
-	}()
-
-	procErrCh := make(chan error, 1)
-	go func() {
-		// os/exec closes StderrPipe from Wait. Draining stderr first avoids
-		// dropping the final FFmpeg failure lines that Health surfaces later.
-		<-scanDone
-		procErrCh <- cmd.Wait()
-	}()
-
-	var procErr error
-	var resultErr error
-	watchdogConsumed := false
-
-	select {
-	case procErr = <-procErrCh:
-		resultErr = procErr
-	case wdErr := <-wdErrCh:
-		watchdogConsumed = true
-		if wdErr != nil {
-			metrics.IncLiveFFmpegStall("watchdog_timeout")
-			a.recordProcessDetail(handle, "transcode stalled - no progress detected")
-			a.Logger.Error().Err(wdErr).Str("sessionId", sessionID).Msg("watchdog triggered process termination")
-			a.terminateProcessGroup(cmd, sessionID)
-			procErr = <-procErrCh
-			resultErr = wdErr
-			break
-		}
-		procErr = <-procErrCh
-		resultErr = procErr
-	case <-parentCtx.Done():
-		a.terminateProcessGroup(cmd, sessionID)
-		procErr = <-procErrCh
-		resultErr = parentCtx.Err()
-	}
-
-	wdCancel()
-	if !watchdogConsumed {
-		if wdErr := <-wdErrCh; wdErr != nil {
-			metrics.IncLiveFFmpegStall("watchdog_timeout")
-			a.recordProcessDetail(handle, "transcode stalled - no progress detected")
-			a.Logger.Error().Err(wdErr).Str("sessionId", sessionID).Msg("watchdog reported failure")
-			if resultErr == nil {
-				resultErr = wdErr
-			}
-		}
-	}
-
-	<-scanDone
-
-	if runtimeFailureLine != "" && (procErr != nil || resultErr != nil) {
-		switch hwBackend {
-		case profiles.GPUBackendVAAPI:
-			a.recordVAAPIRuntimeFailure(sessionID, runtimeFailureLine)
-		case profiles.GPUBackendNVENC:
-			a.recordNVENCRuntimeFailure(sessionID, runtimeFailureLine)
-		}
-	}
-
-	if procErr != nil {
-		a.recordProcessDetail(handle, summarizeProcessExit(procErr))
-		a.Logger.Debug().Err(procErr).Str("sessionId", sessionID).Msg("ffmpeg process exited")
-		return
-	}
-	if resultErr != nil {
-		return
-	}
-	a.clearProcessDetail(handle)
-}
-
-func (a *LocalAdapter) startTimeoutForProfile(sourceType ports.SourceType, profile ports.ProfileSpec) time.Duration {
-	timeout := a.StartTimeout
-	if timeout <= 0 {
-		return timeout
-	}
-	if sourceType == ports.SourceFile || !profile.TranscodeVideo {
-		return timeout
-	}
-	if strings.TrimSpace(profile.HWAccel) != "" {
-		return timeout
-	}
-
-	overrideFloor := 30 * time.Second
-	if profile.EffectiveRuntimeMode == ports.RuntimeModeHQ50 {
-		overrideFloor = 60 * time.Second
-	} else {
-		normalizedProfile := profiles.NormalizeRequestedProfileID(profile.Name)
-		if normalizedProfile != profiles.ProfileSafari && normalizedProfile != profiles.ProfileSafariRuntimeHQ {
-			return timeout
-		}
-	}
-
-	overrideMs := envIntBounded(
-		"XG2G_SAFARI_CPU_START_TIMEOUT_MS",
-		int(overrideFloor/time.Millisecond),
-		int(timeout/time.Millisecond),
-		120000,
-	)
-	override := time.Duration(overrideMs) * time.Millisecond
-	if override > timeout {
-		return override
-	}
-	return timeout
-}
-
-func argsHardwareBackend(args []string) profiles.GPUBackend {
-	for i := range args {
-		if args[i] == "-vaapi_device" {
-			return profiles.GPUBackendVAAPI
-		}
-		switch args[i] {
-		case "h264_nvenc", "hevc_nvenc", "av1_nvenc":
-			return profiles.GPUBackendNVENC
-		}
-	}
-	return profiles.GPUBackendNone
-}
-
-func (a *LocalAdapter) writeFirstFrameMarker(sessionID string) {
-	if !ports.IsSafeSessionID(sessionID) {
-		return
-	}
-	markerPath := ports.SessionFirstFrameMarkerPath(a.HLSRoot, sessionID)
-	if markerPath == "" {
-		return
-	}
-	if err := os.MkdirAll(filepath.Dir(markerPath), 0o750); err != nil {
-		a.Logger.Warn().
-			Err(err).
-			Str("session_id", sessionID).
-			Str("marker_path", markerPath).
-			Msg("failed to prepare first-frame marker directory")
-		return
-	}
-	if err := os.WriteFile(markerPath, []byte(time.Now().UTC().Format(time.RFC3339Nano)), 0o600); err != nil {
-		a.Logger.Warn().
-			Err(err).
-			Str("session_id", sessionID).
-			Str("marker_path", markerPath).
-			Msg("failed to write first-frame marker")
-	}
-}
-
-func scanFFmpegLogTokens(data []byte, atEOF bool) (advance int, token []byte, err error) {
-	if len(data) == 0 {
-		if atEOF {
-			return 0, nil, nil
-		}
-		return 0, nil, nil
-	}
-	for i, b := range data {
-		if b != '\n' && b != '\r' {
-			continue
-		}
-		advance = i + 1
-		if b == '\r' && advance < len(data) && data[advance] == '\n' {
-			advance++
-		}
-		return advance, bytes.TrimRight(data[:i], "\r\n"), nil
-	}
-	if atEOF {
-		return len(data), bytes.TrimRight(data, "\r\n"), nil
-	}
-	return 0, nil, nil
-}
-
-func parseFFmpegFrameCount(line string) (int, bool) {
-	_, after, ok := strings.Cut(line, "frame=")
-	if !ok {
-		return 0, false
-	}
-	rest := strings.TrimLeft(after, " ")
-	if rest == "" {
-		return 0, false
-	}
-	count := 0
-	digits := 0
-	for digits < len(rest) {
-		ch := rest[digits]
-		if ch < '0' || ch > '9' {
-			break
-		}
-		count = count*10 + int(ch-'0')
-		digits++
-	}
-	if digits == 0 {
-		return 0, false
-	}
-	return count, true
-}
-
-func (a *LocalAdapter) recordRuntimeDiagnostics(handle ports.RunHandle, rawLine string, sanitizedLine string) {
-	if handle == "" {
-		return
-	}
-
-	nowUnix := time.Now().Unix()
-	frame, hasFrame := parseFFmpegFrameCount(rawLine)
-	fps, hasFPS := parseFFmpegFloatValue(rawLine, "fps")
-	drops, hasDrops := parseFFmpegIntValue(rawLine, "drop_frames")
-	if !hasDrops {
-		drops, hasDrops = parseFFmpegIntValue(rawLine, "drop")
-	}
-	duplicates, hasDuplicates := parseFFmpegIntValue(rawLine, "dup_frames")
-	if !hasDuplicates {
-		duplicates, hasDuplicates = parseFFmpegIntValue(rawLine, "dup")
-	}
-	speed, hasSpeed := parseFFmpegFloatValue(rawLine, "speed")
-	isProgressLine := hasFrame || hasFPS || hasDrops || hasDuplicates || hasSpeed
-
-	var corruptDecodedFrame bool
-	var warningLine string
-	if !isProgressLine {
-		lower := strings.ToLower(strings.TrimSpace(sanitizedLine))
-		if strings.Contains(lower, "corrupt decoded frame") {
-			corruptDecodedFrame = true
-			warningLine = sanitizedLine
-		} else if !isFFmpegProgressLine(lower) && (summarizeFFmpegFailureLine(lower) != "" || looksLikeFFmpegWarning(lower)) {
-			warningLine = sanitizedLine
-		}
-	}
-
-	if !isProgressLine && warningLine == "" {
-		return
-	}
-
-	var changed bool
-	a.mu.Lock()
-	diagnostics := a.runtimeDiagnostics[handle]
-	if hasFrame {
-		diagnostics.FrameCount = frame
-		changed = true
-	}
-	if hasFPS {
-		diagnostics.FPS = fps
-		changed = true
-	}
-	if hasDrops {
-		diagnostics.DropFrames = drops
-		changed = true
-	}
-	if hasDuplicates {
-		diagnostics.DupFrames = duplicates
-		changed = true
-	}
-	if hasSpeed {
-		diagnostics.Speed = speed
-		changed = true
-	}
-	if corruptDecodedFrame {
-		diagnostics.CorruptDecodedFrames++
-	}
-	if warningLine != "" {
-		diagnostics.LastWarning = trimForLog(warningLine, 240)
-		changed = true
-	}
-	if changed {
-		diagnostics.UpdatedAtUnix = nowUnix
-		if a.runtimeDiagnostics == nil {
-			a.runtimeDiagnostics = make(map[ports.RunHandle]ports.RuntimeDiagnostics)
-		}
-		a.runtimeDiagnostics[handle] = diagnostics
-	}
-	a.mu.Unlock()
-}
-
-func parseFFmpegIntValue(line string, key string) (int, bool) {
-	raw, ok := parseFFmpegValue(line, key)
-	if !ok {
-		return 0, false
-	}
-	value, err := strconv.Atoi(raw)
-	if err != nil {
-		return 0, false
-	}
-	return value, true
-}
-
-func parseFFmpegFloatValue(line string, key string) (float64, bool) {
-	raw, ok := parseFFmpegValue(line, key)
-	if !ok {
-		return 0, false
-	}
-	raw = strings.TrimSuffix(raw, "x")
-	value, err := strconv.ParseFloat(raw, 64)
-	if err != nil {
-		return 0, false
-	}
-	return value, true
-}
-
-func parseFFmpegValue(line string, key string) (string, bool) {
-	key = strings.TrimSpace(key)
-	if key == "" {
-		return "", false
-	}
-	idx := strings.Index(line, key+"=")
-	if idx < 0 {
-		return "", false
-	}
-	rest := strings.TrimLeft(line[idx+len(key)+1:], " ")
-	if rest == "" {
-		return "", false
-	}
-	end := 0
-	for end < len(rest) {
-		ch := rest[end]
-		if (ch >= '0' && ch <= '9') || ch == '.' || ch == '-' || ch == '+' || ch == 'x' {
-			end++
-			continue
-		}
-		break
-	}
-	if end == 0 {
-		return "", false
-	}
-	return rest[:end], true
-}
-
-func extractStartupSegmentPath(line string) (string, bool) {
-	if !strings.Contains(line, "Opening ") || !strings.Contains(line, " for writing") {
-		return "", false
-	}
-	start := strings.IndexAny(line, `'"`)
-	if start < 0 || start+1 >= len(line) {
-		return "", false
-	}
-	quote := line[start]
-	endRel := strings.IndexByte(line[start+1:], quote)
-	if endRel < 0 {
-		return "", false
-	}
-	path := line[start+1 : start+1+endRel]
-	base := filepath.Base(path)
-	if !strings.HasPrefix(base, "seg_") {
-		return "", false
-	}
-	if strings.HasSuffix(base, ".m4s") || strings.HasSuffix(base, ".ts") {
-		return path, true
-	}
-	return "", false
-}
-
-// Stop terminates the process.
-func (a *LocalAdapter) Stop(ctx context.Context, handle ports.RunHandle) error {
-	a.mu.Lock()
-	cmd, exists := a.activeProcs[handle]
-	if exists {
-		a.removeActiveProcessLocked(handle, false)
-	}
-	a.mu.Unlock()
-
-	if !exists {
-		return nil // Idempotent
-	}
-
-	if cmd.Process != nil {
-		return a.killProcessGroup(cmd)
-	}
-
-	return nil
-}
-
-func (a *LocalAdapter) removeActiveProcessLocked(handle ports.RunHandle, archiveDetail bool) {
-	delete(a.activeProcs, handle)
-	delete(a.finalizedProfiles, handle)
-	delete(a.executedPlans, handle)
-	delete(a.runtimeDiagnostics, handle)
-	if archiveDetail {
-		a.archiveProcessDetailLocked(handle)
-	}
-	delete(a.processDetails, handle)
-}
-
-func (a *LocalAdapter) archiveProcessDetailLocked(handle ports.RunHandle) {
-	detail := strings.TrimSpace(a.processDetails[handle])
-	if detail == "" {
-		return
-	}
-	if a.completedProcessDetails == nil {
-		a.completedProcessDetails = make(map[ports.RunHandle]string)
-	}
-	if _, exists := a.completedProcessDetails[handle]; !exists {
-		a.completedProcessDetailOrder = append(a.completedProcessDetailOrder, handle)
-	}
-	a.completedProcessDetails[handle] = detail
-	for len(a.completedProcessDetailOrder) > maxCompletedProcessDetails {
-		evict := a.completedProcessDetailOrder[0]
-		a.completedProcessDetailOrder = a.completedProcessDetailOrder[1:]
-		delete(a.completedProcessDetails, evict)
-	}
-}
-
-func (a *LocalAdapter) terminateProcessGroup(cmd *exec.Cmd, sessionID string) {
-	if err := a.killProcessGroup(cmd); err != nil {
-		a.Logger.Warn().Err(err).Str("sessionId", sessionID).Msg("failed to terminate ffmpeg process group")
-	}
-}
-
-func (a *LocalAdapter) killProcessGroup(cmd *exec.Cmd) error {
-	if cmd == nil || cmd.Process == nil {
-		return nil
-	}
-	// Non-reaping group ladder: the monitor goroutine's cmd.Wait is the sole
-	// reaper, so this must not call Wait or it would race the exit status
-	// (corrupting exit classification, e.g. a spurious GPU->CPU demotion).
-	return procgroup.KillGroupGraceful(cmd.Process.Pid, 2*time.Second, a.KillTimeout)
-}
-
-func (a *LocalAdapter) Health(ctx context.Context, handle ports.RunHandle) ports.HealthStatus {
-	a.mu.Lock()
-	_, exists := a.activeProcs[handle]
-	diagnostics := a.runtimeDiagnostics[handle]
-	a.mu.Unlock()
-	if !exists {
-		// monitorProcess has finished — scanner drained, detail is final.
-		return ports.HealthStatus{
-			Healthy:     false,
-			Message:     a.processStatusMessage(handle, "process not found"),
-			LastCheck:   time.Now(),
-			Diagnostics: diagnostics,
-		}
-	}
-
-	return ports.HealthStatus{
-		Healthy:     true,
-		Message:     "process active",
-		LastCheck:   time.Now(),
-		Diagnostics: diagnostics,
-	}
-}
-
-func envIntBounded(key string, defaultValue, minValue, maxValue int) int {
-	raw := strings.TrimSpace(config.ParseString(key, ""))
-	if raw == "" {
-		return defaultValue
-	}
-	n, err := strconv.Atoi(raw)
-	if err != nil {
-		return defaultValue
-	}
-	if n < minValue {
-		return minValue
-	}
-	if n > maxValue {
-		return maxValue
-	}
-	return n
-}
-
-func envFloatBounded(key string, defaultValue, minValue, maxValue float64) float64 {
-	raw := strings.TrimSpace(config.ParseString(key, ""))
-	if raw == "" {
-		return defaultValue
-	}
-	n, err := strconv.ParseFloat(raw, 64)
-	if err != nil {
-		return defaultValue
-	}
-	if n < minValue {
-		return minValue
-	}
-	if n > maxValue {
-		return maxValue
-	}
-	return n
-}
-
-func envBool(key string, defaultValue bool) bool {
-	return config.ParseBool(key, defaultValue)
-}
-
-func (a *LocalAdapter) recordProcessDetail(handle ports.RunHandle, detail string) {
-	detail = strings.TrimSpace(detail)
-	if detail == "" {
-		return
-	}
-	a.mu.Lock()
-	defer a.mu.Unlock()
-	current := a.processDetails[handle]
-	if processDetailPriority(detail) >= processDetailPriority(current) {
-		a.processDetails[handle] = detail
-	}
-}
-
-func (a *LocalAdapter) clearProcessDetail(handle ports.RunHandle) {
-	a.mu.Lock()
-	delete(a.processDetails, handle)
-	a.mu.Unlock()
-}
-
-func (a *LocalAdapter) processStatusMessage(handle ports.RunHandle, fallback string) string {
-	a.mu.Lock()
-	defer a.mu.Unlock()
-	if detail := strings.TrimSpace(a.processDetails[handle]); detail != "" {
-		delete(a.processDetails, handle)
-		return detail
-	}
-	if detail := strings.TrimSpace(a.completedProcessDetails[handle]); detail != "" {
-		delete(a.completedProcessDetails, handle)
-		return detail
-	}
-	return fallback
-}
-
-func processDetailPriority(detail string) int {
-	switch detail {
-	case "runtime path correctness failed - black output detected":
-		return 55
-	case "transcode stalled - no progress detected":
-		return 50
-	case "copy output missing codec parameters":
-		return 45
-	case "upstream stream ended prematurely":
-		return 40
-	case "failed to open upstream input", "upstream input/output error":
-		return 30
-	case "invalid upstream input data":
-		return 20
-	case "process exited unexpectedly":
-		return 10
-	default:
-		return 0
-	}
-}
-
-func summarizeFFmpegFailureLine(line string) string {
-	lower := strings.ToLower(strings.TrimSpace(line))
-	switch {
-	case strings.Contains(lower, "non-existing pps"),
-		strings.Contains(lower, "non-existing sps"),
-		strings.Contains(lower, "could not find codec parameters for stream") && strings.Contains(lower, "unspecified size"),
-		strings.Contains(lower, "could not write header (incorrect codec parameters ?)"),
-		strings.Contains(lower, "dimensions not set"):
-		return "copy output missing codec parameters"
-	case strings.Contains(lower, "stream ends prematurely"):
-		return "upstream stream ended prematurely"
-	case strings.Contains(lower, "error opening input files"),
-		strings.Contains(lower, "error opening input file"),
-		strings.Contains(lower, "error opening input:"):
-		if strings.Contains(lower, "input/output error") {
-			return "upstream input/output error"
-		}
-		return "failed to open upstream input"
-	case strings.Contains(lower, "invalid data found when processing input"):
-		return "invalid upstream input data"
-	}
-	return ""
-}
-
-func isVAAPIRuntimeFailureLine(line string) bool {
-	lower := strings.ToLower(strings.TrimSpace(line))
-	if lower == "" {
-		return false
-	}
-	if strings.Contains(lower, "vaapi") && looksLikeFFmpegWarning(lower) {
-		return true
-	}
-	definitiveKeywords := []string{
-		"libva error",
-		"no usable encoding entrypoint",
-		"failed to end picture",
-		"failed to sync surface",
-		"failed to export surface",
-		"failed to upload",
-		"hardware device reference is required",
-		"va_create",
-	}
-	for _, keyword := range definitiveKeywords {
-		if strings.Contains(lower, keyword) {
-			return true
-		}
-	}
-	if (strings.Contains(lower, "hwupload") || strings.Contains(lower, "renderd128")) && looksLikeFFmpegWarning(lower) {
-		return true
-	}
-	return false
-}
-
-func isNVENCRuntimeFailureLine(line string) bool {
-	lower := strings.ToLower(strings.TrimSpace(line))
-	if lower == "" {
-		return false
-	}
-	if strings.Contains(lower, "nvenc") && looksLikeFFmpegWarning(lower) {
-		return true
-	}
-	definitiveKeywords := []string{
-		"cannot load libnvidia-encode.so.1",
-		"driver does not support the required nvenc api version",
-		"no capable devices found",
-		"no nvenc capable devices found",
-		"openencode session failed",
-		"provided device doesn't support required nvenc features",
-		"cannot init encoder",
-		"nvidia",
-	}
-	for _, keyword := range definitiveKeywords {
-		if strings.Contains(lower, keyword) && looksLikeFFmpegWarning(lower) {
-			return true
-		}
-	}
-	return false
-}
-
-func (a *LocalAdapter) recordVAAPIRuntimeFailure(sessionID, failureLine string) {
-	if !hardware.IsVAAPIReady() {
-		return
-	}
-	failures, demoted := hardware.RecordVAAPIRuntimeFailure()
-	event := a.Logger.Warn().
-		Str("session_id", sessionID).
-		Int("vaapi_runtime_failures", failures)
-	if failureLine != "" {
-		event = event.Str("ffmpeg_log", failureLine)
-	}
-	if demoted {
-		event.Msg("vaapi runtime failure threshold reached; gpu demoted to cpu fallback")
-		return
-	}
-	event.Msg("recorded vaapi runtime failure")
-}
-
-func (a *LocalAdapter) recordNVENCRuntimeFailure(sessionID, failureLine string) {
-	if !hardware.IsNVENCReady() {
-		return
-	}
-	failures, demoted := hardware.RecordNVENCRuntimeFailure()
-	event := a.Logger.Warn().
-		Str("session_id", sessionID).
-		Int("nvenc_runtime_failures", failures)
-	if failureLine != "" {
-		event = event.Str("ffmpeg_log", failureLine)
-	}
-	if demoted {
-		event.Msg("nvenc runtime failure threshold reached; gpu demoted to cpu fallback")
-		return
-	}
-	event.Msg("recorded nvenc runtime failure")
-}
-
-func ffmpegLogLevel(line string) zerolog.Level {
-	lower := strings.ToLower(strings.TrimSpace(line))
-	if lower == "" {
-		return zerolog.DebugLevel
-	}
-	if isFFmpegProgressLine(lower) {
-		return zerolog.DebugLevel
-	}
-	if summarizeFFmpegFailureLine(lower) != "" || looksLikeFFmpegWarning(lower) {
-		return zerolog.WarnLevel
-	}
-	return zerolog.InfoLevel
-}
-
-func isFFmpegProgressLine(lower string) bool {
-	switch {
-	case strings.HasPrefix(lower, "frame="),
-		strings.HasPrefix(lower, "fps="),
-		strings.HasPrefix(lower, "stream_"),
-		strings.HasPrefix(lower, "bitrate="),
-		strings.HasPrefix(lower, "total_size="),
-		strings.HasPrefix(lower, "out_time_us="),
-		strings.HasPrefix(lower, "out_time_ms="),
-		strings.HasPrefix(lower, "out_time="),
-		strings.HasPrefix(lower, "dup_frames="),
-		strings.HasPrefix(lower, "drop_frames="),
-		strings.HasPrefix(lower, "speed="),
-		strings.HasPrefix(lower, "progress="):
-		return true
-	case strings.Contains(lower, " opening '") && strings.Contains(lower, "' for writing"):
-		return true
-	case strings.Contains(lower, "opening \"") && strings.Contains(lower, "\" for writing"):
-		return true
-	case strings.Contains(lower, "press [q] to stop"):
-		return true
-	default:
-		return false
-	}
-}
-
-func looksLikeFFmpegWarning(lower string) bool {
-	keywords := []string{
-		" error",
-		"error ",
-		"failed",
-		"invalid",
-		"non-existing",
-		"no frame",
-		"decode_slice_header",
-		"corrupt",
-		"unable to",
-		"could not",
-		"connection refused",
-		"broken pipe",
-	}
-	for _, keyword := range keywords {
-		if strings.Contains(lower, keyword) {
-			return true
-		}
-	}
-	return false
-}
-
-func summarizeProcessExit(procErr error) string {
-	if procErr == nil {
-		return ""
-	}
-	var exitErr *exec.ExitError
-	if errors.As(procErr, &exitErr) {
-		return fmt.Sprintf("process exit code %d", exitErr.ExitCode())
-	}
-	return "process exited unexpectedly"
-}
-
-// Preflight delegators keep the adapter's external surface (bootstrap wiring)
-// stable while encoder-capability detection lives in the Detector.
-func (a *LocalAdapter) PreflightVAAPI() error       { return a.detector.PreflightVAAPI() }
-func (a *LocalAdapter) PreflightNVENC() error       { return a.detector.PreflightNVENC() }
-func (a *LocalAdapter) PreflightTranscodeProfiles() { a.detector.PreflightTranscodeProfiles() }
-
-func (a *LocalAdapter) PreflightPathCorrectness() { a.detector.PreflightPathCorrectness() }
