+++ backend/internal/infra/media/ffmpeg/adapter_process.go
@@ -1,0 +1,641 @@
+package ffmpeg
+
+import (
+	"bufio"
+	"context"
+	"fmt"
+	"github.com/ManuGH/xg2g/internal/config"
+	"github.com/ManuGH/xg2g/internal/domain/session/ports"
+	"github.com/ManuGH/xg2g/internal/infra/ffmpeg/watchdog"
+	"github.com/ManuGH/xg2g/internal/metrics"
+	"github.com/ManuGH/xg2g/internal/pipeline/exec/enigma2"
+	"github.com/ManuGH/xg2g/internal/pipeline/profiles"
+	"github.com/ManuGH/xg2g/internal/procgroup"
+	"github.com/ManuGH/xg2g/internal/telemetry"
+	"github.com/rs/zerolog"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/codes"
+	"go.opentelemetry.io/otel/trace"
+	"io"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strconv"
+	"strings"
+	"sync"
+	"sync/atomic"
+	"time"
+)
+
+// Start initiates the media process.
+func (a *LocalAdapter) Start(ctx context.Context, spec ports.StreamSpec) (ports.RunHandle, error) {
+	if spec.Source.Type == ports.SourceTuner && a.E2 != nil {
+		if spec.Source.TunerSlot < 0 {
+			return "", fmt.Errorf("invalid tuner slot: %d", spec.Source.TunerSlot)
+		}
+		tuner := enigma2.NewTuner(a.E2, spec.Source.TunerSlot, 10*time.Second)
+		if err := tuner.Tune(ctx, spec.Source.ID); err != nil {
+			return "", fmt.Errorf("tuning failed: %w", err)
+		}
+		a.Logger.Info().
+			Str("session_id", spec.SessionID).
+			Str("startup_phase", "tuner_tuned").
+			Int("tuner_slot", spec.Source.TunerSlot).
+			Str("service_ref", spec.Source.ID).
+			Msg("tuner tune completed")
+	}
+
+	inputURL := ""
+	switch spec.Source.Type {
+	case ports.SourceTuner:
+		if a.E2 == nil {
+			return "", fmt.Errorf("tuner source requires enigma2 client")
+		}
+		streamURL, err := a.E2.ResolveStreamURL(ctx, spec.Source.ID)
+		if err != nil {
+			return "", fmt.Errorf("resolve stream url: %w", err)
+		}
+		streamURL = a.injectCredentialsIfAllowed(streamURL)
+		a.Logger.Info().
+			Str("session_id", spec.SessionID).
+			Str("startup_phase", "stream_url_resolved").
+			Str("resolved_url", sanitizeURLForLog(streamURL)).
+			Msg("stream url resolved")
+
+		chosenURL, err := a.selectStreamURL(ctx, spec.SessionID, spec.Source.ID, streamURL)
+		if err != nil {
+			return "", err
+		}
+		inputURL = chosenURL
+		a.Logger.Info().
+			Str("session_id", spec.SessionID).
+			Str("startup_phase", "input_url_selected").
+			Str("input_url", sanitizeURLForLog(inputURL)).
+			Msg("stream input url selected")
+	case ports.SourceURL:
+		inputURL = spec.Source.ID
+		a.Logger.Info().
+			Str("session_id", spec.SessionID).
+			Str("startup_phase", "input_url_selected").
+			Str("input_url", sanitizeURLForLog(inputURL)).
+			Msg("direct input url selected")
+	}
+	sourceKey := fpsCacheKey(spec.Source, inputURL)
+
+	a.Logger.Info().
+		Str("session_id", spec.SessionID).
+		Str("startup_phase", "ffmpeg_args_build_started").
+		Str("input_url", sanitizeURLForLog(inputURL)).
+		Msg("ffmpeg args build started")
+	// Span covering the playback decision + ffprobe probes (the untraced gap
+	// between request and spawn). No-op when tracing is off. planCtx lets any
+	// probe spans added later nest under this one.
+	planCtx, planSpan := telemetry.Tracer("xg2g.ffmpeg").Start(ctx, "playback.plan",
+		trace.WithAttributes(
+			attribute.String("xg2g.session_id", spec.SessionID),
+			attribute.String("xg2g.source_type", fmt.Sprintf("%v", spec.Source.Type)),
+		),
+	)
+	plan, err := a.buildArgsWithPlan(planCtx, spec, inputURL)
+	if err != nil {
+		planSpan.RecordError(err)
+		planSpan.SetStatus(codes.Error, "plan build failed")
+		planSpan.End()
+		return "", fmt.Errorf("failed to build args: %w", err)
+	}
+	planSpan.SetAttributes(attribute.String("xg2g.path_id", plan.pathID))
+	planSpan.End()
+	args := plan.args
+	a.Logger.Info().
+		Str("session_id", spec.SessionID).
+		Str("startup_phase", "ffmpeg_args_built").
+		Int("arg_count", len(args)).
+		Msg("ffmpeg args build finished")
+
+	// #nosec G204 - BinPath is trusted from config; args are generated by strict internal logic (buildArgs)
+	cmd := exec.CommandContext(ctx, a.BinPath, args...)
+	procgroup.Set(cmd) // Mandatory for tree reaping
+	// On ctx cancellation, gracefully SIGTERM the whole process group so ffmpeg
+	// can flush its final segment + ENDLIST, instead of the stdlib default which
+	// SIGKILLs the leader PID only. WaitDelay then bounds how long Wait blocks
+	// after cancellation before the runtime force-kills the leader and closes I/O.
+	cmd.Cancel = func() error {
+		if cmd.Process == nil {
+			return nil
+		}
+		return procgroup.TerminateGroup(cmd.Process.Pid)
+	}
+	if a.KillTimeout > 0 {
+		cmd.WaitDelay = a.KillTimeout
+	} else {
+		cmd.WaitDelay = 5 * time.Second
+	}
+
+	stderr, err := cmd.StderrPipe()
+	if err != nil {
+		return "", fmt.Errorf("failed to pipe stderr: %w", err)
+	}
+	cmd.Stdout = nil
+
+	// Span covering ffmpeg spawn -> first HLS segment (the user-perceived startup
+	// latency). No-op when tracing is disabled; ended exactly once in the monitor.
+	spawnedAt := time.Now()
+	_, startupSpan := telemetry.Tracer("xg2g.ffmpeg").Start(ctx, "ffmpeg.startup",
+		trace.WithAttributes(
+			attribute.String("xg2g.session_id", spec.SessionID),
+			attribute.String("xg2g.source_type", fmt.Sprintf("%v", spec.Source.Type)),
+			attribute.String("xg2g.hw_backend", fmt.Sprintf("%v", argsHardwareBackend(args))),
+			attribute.String("xg2g.path_id", plan.pathID),
+		),
+	)
+
+	if err := cmd.Start(); err != nil {
+		startupSpan.RecordError(err)
+		startupSpan.SetStatus(codes.Error, "ffmpeg start failed")
+		startupSpan.End()
+		return "", fmt.Errorf("ffmpeg start failed: %w", err)
+	}
+
+	pid := cmd.Process.Pid
+	a.Logger.Info().
+		Str("session_id", spec.SessionID).
+		Str("startup_phase", "ffmpeg_started").
+		Int("pid", pid).
+		Msg("ffmpeg process started")
+	handle := ports.RunHandle(fmt.Sprintf("%s-%d", spec.SessionID, pid))
+	a.mu.Lock()
+	a.activeProcs[handle] = cmd
+	a.finalizedProfiles[handle] = plan.effectiveProfile
+	// Execution truth: capture the plan parsed from the REAL argv we just handed
+	// to the process, so observers report what ffmpeg runs, not a prediction.
+	a.executedPlans[handle] = executedFFmpegPlanFromArgs(args)
+	delete(a.processDetails, handle)
+	a.mu.Unlock()
+
+	go a.monitorProcessWithStartTimeout(ctx, handle, cmd, stderr, spec.SessionID, argsHardwareBackend(args), plan.pathID, a.startTimeoutForProfile(spec.Source.Type, plan.effectiveProfile), startupSpan, spawnedAt) // #nosec G118 -- goroutine receives the request-scoped ctx (first arg), not context.Background/TODO
+	if sourceKey != "" {
+		go a.learnFPSFromOutput(sourceKey, spec.SessionID)
+	}
+
+	metrics.RecordPipelineSpawn("ffmpeg", "admitted")
+	return handle, nil
+}
+
+func (a *LocalAdapter) FinalizedProfile(handle ports.RunHandle) (ports.ProfileSpec, bool) {
+	a.mu.Lock()
+	defer a.mu.Unlock()
+
+	profile, ok := a.finalizedProfiles[handle]
+	if !ok {
+		return ports.ProfileSpec{}, false
+	}
+	return profile, true
+}
+
+// ExecutedFFmpegPlan returns the execution-truth plan parsed from the real argv
+// of the process launched for the handle. It is the un-lie-able source for
+// "what ffmpeg runs", as opposed to any profile-derived prediction.
+func (a *LocalAdapter) ExecutedFFmpegPlan(handle ports.RunHandle) (ports.ExecutedFFmpegPlan, bool) {
+	a.mu.Lock()
+	defer a.mu.Unlock()
+
+	plan, ok := a.executedPlans[handle]
+	if !ok {
+		return ports.ExecutedFFmpegPlan{}, false
+	}
+	return plan, true
+}
+
+func (a *LocalAdapter) monitorProcessWithStartTimeout(parentCtx context.Context, handle ports.RunHandle, cmd *exec.Cmd, stderr io.ReadCloser, sessionID string, hwBackend profiles.GPUBackend, pathID string, startTimeout time.Duration, startupSpan trace.Span, spawnedAt time.Time) {
+	defer func() {
+		a.mu.Lock()
+		a.removeActiveProcessLocked(handle, true)
+		a.mu.Unlock()
+	}()
+
+	// End the startup span exactly once: success is recorded on first segment;
+	// any other exit path (start timeout, early ffmpeg exit, ctx cancel) is an error.
+	var sawFirstSegment atomic.Bool
+	endStartupSpan := sync.OnceFunc(func() { startupSpan.End() })
+	defer func() {
+		if !sawFirstSegment.Load() {
+			startupSpan.SetStatus(codes.Error, "ffmpeg exited before first HLS segment")
+		}
+		endStartupSpan()
+	}()
+
+	wd := watchdog.New(startTimeout, a.StallTimeout)
+
+	if parentCtx == nil {
+		parentCtx = context.Background()
+	}
+	wdCtx, wdCancel := context.WithCancel(parentCtx)
+	defer wdCancel()
+	observerCtx, observerCancel := context.WithCancel(parentCtx)
+	defer observerCancel()
+
+	wdErrCh := make(chan error, 1)
+	go func() {
+		wdErrCh <- wd.Run(wdCtx)
+	}()
+
+	runtimeFailureLine := ""
+	scanDone := make(chan struct{})
+	go func() {
+		defer close(scanDone)
+
+		// Forward lines to RingBuffer/Log and Watchdog.
+		scanner := bufio.NewScanner(stderr)
+		scanner.Split(scanFFmpegLogTokens)
+		firstFrameLogged := false
+		firstSegmentLogged := false
+		outputObserverStarted := false
+
+		for scanner.Scan() {
+			line := scanner.Text()
+			if !firstFrameLogged {
+				if frame, ok := parseFFmpegFrameCount(line); ok && frame > 0 {
+					firstFrameLogged = true
+					wd.ObserveProgress()
+					a.writeFirstFrameMarker(sessionID)
+					a.Logger.Info().
+						Str("session_id", sessionID).
+						Str("startup_phase", "first_frame").
+						Int("frame", frame).
+						Msg("ffmpeg first frame observed")
+				}
+			}
+			if !firstSegmentLogged {
+				if segmentPath, ok := extractStartupSegmentPath(line); ok {
+					firstSegmentLogged = true
+					sawFirstSegment.Store(true)
+					startupSpan.SetAttributes(attribute.Int64("xg2g.time_to_first_segment_ms", time.Since(spawnedAt).Milliseconds()))
+					startupSpan.SetStatus(codes.Ok, "")
+					endStartupSpan()
+					wd.ObserveProgress()
+					a.Logger.Info().
+						Str("session_id", sessionID).
+						Str("startup_phase", "first_segment_write").
+						Str("segment_path", segmentPath).
+						Msg("ffmpeg first segment write observed")
+					if pathID != "" && !outputObserverStarted {
+						outputObserverStarted = true
+						go a.detector.observeRuntimePathCorrectness(observerCtx, handle, cmd, sessionID, pathID)
+					}
+				}
+			}
+			sanitizedLine := sanitizeFFmpegLogLine(line)
+			a.recordRuntimeDiagnostics(handle, line, sanitizedLine)
+			if detail := summarizeFFmpegFailureLine(sanitizedLine); detail != "" {
+				a.recordProcessDetail(handle, detail)
+			}
+			if runtimeFailureLine == "" {
+				switch hwBackend {
+				case profiles.GPUBackendVAAPI:
+					if isVAAPIRuntimeFailureLine(sanitizedLine) {
+						runtimeFailureLine = sanitizedLine
+					}
+				case profiles.GPUBackendNVENC:
+					if isNVENCRuntimeFailureLine(sanitizedLine) {
+						runtimeFailureLine = sanitizedLine
+					}
+				}
+			}
+			switch ffmpegLogLevel(sanitizedLine) {
+			case zerolog.WarnLevel:
+				a.Logger.Warn().Str("sessionId", sessionID).Str("ffmpeg_log", sanitizedLine).Msg("ffmpeg output")
+			case zerolog.InfoLevel:
+				a.Logger.Info().Str("sessionId", sessionID).Str("ffmpeg_log", sanitizedLine).Msg("ffmpeg output")
+			default:
+				a.Logger.Debug().Str("sessionId", sessionID).Str("ffmpeg_log", sanitizedLine).Msg("ffmpeg output")
+			}
+			wd.ParseLine(line)
+		}
+		if scanErr := scanner.Err(); scanErr != nil {
+			a.Logger.Warn().Err(scanErr).Str("sessionId", sessionID).Msg("ffmpeg stderr scan failed")
+		}
+	}()
+
+	procErrCh := make(chan error, 1)
+	go func() {
+		// os/exec closes StderrPipe from Wait. Draining stderr first avoids
+		// dropping the final FFmpeg failure lines that Health surfaces later.
+		<-scanDone
+		procErrCh <- cmd.Wait()
+	}()
+
+	var procErr error
+	var resultErr error
+	watchdogConsumed := false
+
+	select {
+	case procErr = <-procErrCh:
+		resultErr = procErr
+	case wdErr := <-wdErrCh:
+		watchdogConsumed = true
+		if wdErr != nil {
+			metrics.IncLiveFFmpegStall("watchdog_timeout")
+			a.recordProcessDetail(handle, "transcode stalled - no progress detected")
+			a.Logger.Error().Err(wdErr).Str("sessionId", sessionID).Msg("watchdog triggered process termination")
+			a.terminateProcessGroup(cmd, sessionID)
+			procErr = <-procErrCh
+			resultErr = wdErr
+			break
+		}
+		procErr = <-procErrCh
+		resultErr = procErr
+	case <-parentCtx.Done():
+		a.terminateProcessGroup(cmd, sessionID)
+		procErr = <-procErrCh
+		resultErr = parentCtx.Err()
+	}
+
+	wdCancel()
+	if !watchdogConsumed {
+		if wdErr := <-wdErrCh; wdErr != nil {
+			metrics.IncLiveFFmpegStall("watchdog_timeout")
+			a.recordProcessDetail(handle, "transcode stalled - no progress detected")
+			a.Logger.Error().Err(wdErr).Str("sessionId", sessionID).Msg("watchdog reported failure")
+			if resultErr == nil {
+				resultErr = wdErr
+			}
+		}
+	}
+
+	<-scanDone
+
+	if runtimeFailureLine != "" && (procErr != nil || resultErr != nil) {
+		switch hwBackend {
+		case profiles.GPUBackendVAAPI:
+			a.recordVAAPIRuntimeFailure(sessionID, runtimeFailureLine)
+		case profiles.GPUBackendNVENC:
+			a.recordNVENCRuntimeFailure(sessionID, runtimeFailureLine)
+		}
+	}
+
+	if procErr != nil {
+		a.recordProcessDetail(handle, summarizeProcessExit(procErr))
+		a.Logger.Debug().Err(procErr).Str("sessionId", sessionID).Msg("ffmpeg process exited")
+		return
+	}
+	if resultErr != nil {
+		return
+	}
+	a.clearProcessDetail(handle)
+}
+
+func (a *LocalAdapter) startTimeoutForProfile(sourceType ports.SourceType, profile ports.ProfileSpec) time.Duration {
+	timeout := a.StartTimeout
+	if timeout <= 0 {
+		return timeout
+	}
+	if sourceType == ports.SourceFile || !profile.TranscodeVideo {
+		return timeout
+	}
+	if strings.TrimSpace(profile.HWAccel) != "" {
+		return timeout
+	}
+
+	overrideFloor := 30 * time.Second
+	if profile.EffectiveRuntimeMode == ports.RuntimeModeHQ50 {
+		overrideFloor = 60 * time.Second
+	} else {
+		normalizedProfile := profiles.NormalizeRequestedProfileID(profile.Name)
+		if normalizedProfile != profiles.ProfileSafari && normalizedProfile != profiles.ProfileSafariRuntimeHQ {
+			return timeout
+		}
+	}
+
+	overrideMs := envIntBounded(
+		"XG2G_SAFARI_CPU_START_TIMEOUT_MS",
+		int(overrideFloor/time.Millisecond),
+		int(timeout/time.Millisecond),
+		120000,
+	)
+	override := time.Duration(overrideMs) * time.Millisecond
+	if override > timeout {
+		return override
+	}
+	return timeout
+}
+
+func argsHardwareBackend(args []string) profiles.GPUBackend {
+	for i := range args {
+		if args[i] == "-vaapi_device" {
+			return profiles.GPUBackendVAAPI
+		}
+		switch args[i] {
+		case "h264_nvenc", "hevc_nvenc", "av1_nvenc":
+			return profiles.GPUBackendNVENC
+		}
+	}
+	return profiles.GPUBackendNone
+}
+
+func (a *LocalAdapter) writeFirstFrameMarker(sessionID string) {
+	if !ports.IsSafeSessionID(sessionID) {
+		return
+	}
+	markerPath := ports.SessionFirstFrameMarkerPath(a.HLSRoot, sessionID)
+	if markerPath == "" {
+		return
+	}
+	if err := os.MkdirAll(filepath.Dir(markerPath), 0o750); err != nil {
+		a.Logger.Warn().
+			Err(err).
+			Str("session_id", sessionID).
+			Str("marker_path", markerPath).
+			Msg("failed to prepare first-frame marker directory")
+		return
+	}
+	if err := os.WriteFile(markerPath, []byte(time.Now().UTC().Format(time.RFC3339Nano)), 0o600); err != nil {
+		a.Logger.Warn().
+			Err(err).
+			Str("session_id", sessionID).
+			Str("marker_path", markerPath).
+			Msg("failed to write first-frame marker")
+	}
+}
+
+// Stop terminates the process.
+func (a *LocalAdapter) Stop(ctx context.Context, handle ports.RunHandle) error {
+	a.mu.Lock()
+	cmd, exists := a.activeProcs[handle]
+	if exists {
+		a.removeActiveProcessLocked(handle, false)
+	}
+	a.mu.Unlock()
+
+	if !exists {
+		return nil // Idempotent
+	}
+
+	if cmd.Process != nil {
+		return a.killProcessGroup(cmd)
+	}
+
+	return nil
+}
+
+func (a *LocalAdapter) removeActiveProcessLocked(handle ports.RunHandle, archiveDetail bool) {
+	delete(a.activeProcs, handle)
+	delete(a.finalizedProfiles, handle)
+	delete(a.executedPlans, handle)
+	delete(a.runtimeDiagnostics, handle)
+	if archiveDetail {
+		a.archiveProcessDetailLocked(handle)
+	}
+	delete(a.processDetails, handle)
+}
+
+func (a *LocalAdapter) archiveProcessDetailLocked(handle ports.RunHandle) {
+	detail := strings.TrimSpace(a.processDetails[handle])
+	if detail == "" {
+		return
+	}
+	if a.completedProcessDetails == nil {
+		a.completedProcessDetails = make(map[ports.RunHandle]string)
+	}
+	if _, exists := a.completedProcessDetails[handle]; !exists {
+		a.completedProcessDetailOrder = append(a.completedProcessDetailOrder, handle)
+	}
+	a.completedProcessDetails[handle] = detail
+	for len(a.completedProcessDetailOrder) > maxCompletedProcessDetails {
+		evict := a.completedProcessDetailOrder[0]
+		a.completedProcessDetailOrder = a.completedProcessDetailOrder[1:]
+		delete(a.completedProcessDetails, evict)
+	}
+}
+
+func (a *LocalAdapter) terminateProcessGroup(cmd *exec.Cmd, sessionID string) {
+	if err := a.killProcessGroup(cmd); err != nil {
+		a.Logger.Warn().Err(err).Str("sessionId", sessionID).Msg("failed to terminate ffmpeg process group")
+	}
+}
+
+func (a *LocalAdapter) killProcessGroup(cmd *exec.Cmd) error {
+	if cmd == nil || cmd.Process == nil {
+		return nil
+	}
+	// Non-reaping group ladder: the monitor goroutine's cmd.Wait is the sole
+	// reaper, so this must not call Wait or it would race the exit status
+	// (corrupting exit classification, e.g. a spurious GPU->CPU demotion).
+	return procgroup.KillGroupGraceful(cmd.Process.Pid, 2*time.Second, a.KillTimeout)
+}
+
+func (a *LocalAdapter) Health(ctx context.Context, handle ports.RunHandle) ports.HealthStatus {
+	a.mu.Lock()
+	_, exists := a.activeProcs[handle]
+	diagnostics := a.runtimeDiagnostics[handle]
+	a.mu.Unlock()
+	if !exists {
+		// monitorProcess has finished — scanner drained, detail is final.
+		return ports.HealthStatus{
+			Healthy:     false,
+			Message:     a.processStatusMessage(handle, "process not found"),
+			LastCheck:   time.Now(),
+			Diagnostics: diagnostics,
+		}
+	}
+
+	return ports.HealthStatus{
+		Healthy:     true,
+		Message:     "process active",
+		LastCheck:   time.Now(),
+		Diagnostics: diagnostics,
+	}
+}
+
+func envIntBounded(key string, defaultValue, minValue, maxValue int) int {
+	raw := strings.TrimSpace(config.ParseString(key, ""))
+	if raw == "" {
+		return defaultValue
+	}
+	n, err := strconv.Atoi(raw)
+	if err != nil {
+		return defaultValue
+	}
+	if n < minValue {
+		return minValue
+	}
+	if n > maxValue {
+		return maxValue
+	}
+	return n
+}
+
+func envFloatBounded(key string, defaultValue, minValue, maxValue float64) float64 {
+	raw := strings.TrimSpace(config.ParseString(key, ""))
+	if raw == "" {
+		return defaultValue
+	}
+	n, err := strconv.ParseFloat(raw, 64)
+	if err != nil {
+		return defaultValue
+	}
+	if n < minValue {
+		return minValue
+	}
+	if n > maxValue {
+		return maxValue
+	}
+	return n
+}
+
+func envBool(key string, defaultValue bool) bool {
+	return config.ParseBool(key, defaultValue)
+}
+
+func (a *LocalAdapter) recordProcessDetail(handle ports.RunHandle, detail string) {
+	detail = strings.TrimSpace(detail)
+	if detail == "" {
+		return
+	}
+	a.mu.Lock()
+	defer a.mu.Unlock()
+	current := a.processDetails[handle]
+	if processDetailPriority(detail) >= processDetailPriority(current) {
+		a.processDetails[handle] = detail
+	}
+}
+
+func (a *LocalAdapter) clearProcessDetail(handle ports.RunHandle) {
+	a.mu.Lock()
+	delete(a.processDetails, handle)
+	a.mu.Unlock()
+}
+
+func (a *LocalAdapter) processStatusMessage(handle ports.RunHandle, fallback string) string {
+	a.mu.Lock()
+	defer a.mu.Unlock()
+	if detail := strings.TrimSpace(a.processDetails[handle]); detail != "" {
+		delete(a.processDetails, handle)
+		return detail
+	}
+	if detail := strings.TrimSpace(a.completedProcessDetails[handle]); detail != "" {
+		delete(a.completedProcessDetails, handle)
+		return detail
+	}
+	return fallback
+}
+
+func processDetailPriority(detail string) int {
+	switch detail {
+	case "runtime path correctness failed - black output detected":
+		return 55
+	case "transcode stalled - no progress detected":
+		return 50
+	case "copy output missing codec parameters":
+		return 45
+	case "upstream stream ended prematurely":
+		return 40
+	case "failed to open upstream input", "upstream input/output error":
+		return 30
+	case "invalid upstream input data":
+		return 20
+	case "process exited unexpectedly":
+		return 10
+	default:
+		return 0
+	}
+}
